+++ client.js
@@ -1,6 +1,6 @@
-import { DDPLink } from '@swydo/apollo-link-ddp';
+import { DDPLink } from 'meteor-apollo-link-ddp';
 
-export * from '@swydo/apollo-link-ddp';
+export * from 'meteor-apollo-link-ddp';
 
 // It is common for Apollo Links to export the link itself as the default
 export default DDPLink;