+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 4.0.2
 - Bump package versions
-- Replace old imports using apollo-link-ddp to @swydo/apollo-link-ddp
+- Replace old imports using apollo-link-ddp to meteor-apollo-link-ddp
 
 ## 4.0.1
 - Import from @swydo/scoped package